requirements.txt
requests_oauthlib
django-braces